+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] - 2020-11-16
+- [#337](https://github.com/equinor/webviz-config/pull/337) - New generic plugin to
+generate a [Pivot table](https://en.wikipedia.org/wiki/Pivot_table) based on
+[Dash Pivottable](https://github.com/plotly/dash-pivottable).
 
 ## Fixed
 - [#339](https://github.com/equinor/webviz-config/pull/339) - After